chore: migrated next step button to header from footer
